Milestone 4: Generators and Energy Regeneration is implemented.
To test the new features:

Tap the "build" and "agriculture" icons in the AppBar to place a Barracks and a Mine on the grid.
Observe the energy regeneration in the AppBar. The energy should increase by 1 every 60 seconds, up to the maximum.
Tap a Barracks or Mine tile. If there's an adjacent empty tile and enough energy, the generator should spawn its item (sword or coin) and go on cooldown. If there's no empty tile or not enough energy, a message should be printed to the console.
Observe the cooldown state of the generator. After tapping, it should not be immediately tappable again.